Dependencies and configuration are provided on instance base, by dependency injection
This commit addresses issues related to static state. In Devhub we provide dependecies
through dependency injection. Static access is considered a bad practise. This commit
addresses that problem, and provides the `PegDownParser` and `MarkDownParser` by DI.

Concurrent access to the `PegDownParser` is prevented by binding it to the `RequestScope`.
No synchronized access is needed, so we keep our concurrency high.

As the `TemplateEngine` was accessed as a singleton, and it depends on the `PegDownParser`,
access to the `TemplateEngine` had to be transformed to `RequestScoped`, which is done
by using Guice `Providers`.