Implement TemplateDirectiveModule, so noescape and BeansWrapper are not needed
The ideal way to have objects that write to the Freemarker output is to use a directive.
This commit implements such a directive on the existing Markdown parser object.
With this change, `[#noescape]` tags around the Markdown blocks are no longer required.
Furthermore, we do not need the BeansWrapper anymore.