Move existing rubrics to the next course edition
If a course already had a previous edition we add the same rubrics
to the next edition.

Manually generate an assignment id to avoid the ID generation deadlock.
Persist the course edition first then add the assignments.

The courseEdition-create-error was being referenced from a controller,
but not actually present in the translation file.

Also pulled in assertJ for writing tests.

fixes #420