Finish the angular implementation for rubric migration.
Use ng-options for the select boxes, this allows putting
the id as value in and getting the selected object in ng-model.

If there is nothing selected the default value will be a ?,
When we try to turn this into a number we will fail and not
copy any assignments over.